backend/requirements.txt
Django==2.2.19
importlib-metadata==4.2.0
mccabe==0.7.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.3.0
zipp==3.8.1
Pillow==9.2.0
sorl-thumbnail==12.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-filter==22.1
